Be less like MySQL
After all, it was a bad idea to try to be too clever with MERGE or ON
CONFLICT … UPDATE.  This would be better to handle with a MERGE or ON
CONFLICT … UPDATE in the context that's _working_ with the view.  Also,
now we will explicitly pass through overridden GENERATED AS primary
keys, to make sure that we crash instead of doing a MySQL-esque INSERT
that ignores the explicit value.