`Accept-Language` interpretation is best left to another layer
It would be difficult (probably impossible) for
`pg_xenophile_user_lang_code()` to make the right choices for all the
use cases based on parsing the `Accept-Language` header (passed on by
PostgREST in the `request.headers` settings) itself.  Maybe, instead,
I will, in a later version, offer a function to parse this header to
make it a bit easier for the extension user to decide.